refactor question-set create pages
