+++ apps/hub/src/app/(frontpage)/question-sets/LayoutActions.tsx
@@ -2,44 +2,24 @@
 import { useRouter } from "next/navigation";
 import { FC } from "react";
 
-import { ButtonWithDropdown, DropdownMenu, PlusIcon } from "@quri/ui";
+import { Button, PlusIcon } from "@quri/ui";
 
-import { DropdownMenuNextLinkItem } from "@/components/ui/DropdownMenuNextLinkItem";
-import {
-  createQuestionSetFromGitHubIssuesRoute,
-  createQuestionSetFromMetaforecastRoute,
-  createQuestionSetRoute,
-} from "@/lib/routes";
+import { createQuestionSetRoute } from "@/lib/routes";
 
 export const LayoutActions: FC = () => {
   const router = useRouter();
 
   return (
-    <ButtonWithDropdown
-      renderDropdown={() => (
-        <DropdownMenu>
-          <DropdownMenuNextLinkItem
-            href={createQuestionSetRoute()}
-            title="Manually"
-          />
-          <DropdownMenuNextLinkItem
-            href={createQuestionSetFromMetaforecastRoute()}
-            title="From Metaforecast"
-          />
-          <DropdownMenuNextLinkItem
-            href={createQuestionSetFromGitHubIssuesRoute()}
-            title="From GitHub Issues"
-          />
-        </DropdownMenu>
-      )}
+    <Button
       onClick={() => {
         router.push(createQuestionSetRoute());
       }}
+      theme="primary"
     >
       <div className="flex items-center gap-1">
         <PlusIcon size={16} />
         New Question Set
       </div>
-    </ButtonWithDropdown>
+    </Button>
   );
 };
