--- apps/hub/src/app/question-sets/create-from-metaforecast/layout.tsx
@@ -1,9 +1,0 @@
-import { NarrowPageLayout } from "@/components/layout/NarrowPageLayout";
-
-export default function CreateQuestionSetFromMetaforecastLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <NarrowPageLayout>{children}</NarrowPageLayout>;
-}
