+++ apps/hub/src/app/question-sets/create/metaforecast/page.tsx
@@ -1,0 +1,9 @@
+import { loadTopBinaryQuestions } from "@/metaforecast-questions/data/manifold-questions";
+
+import { CreateFromMetaforecastForm } from "./CreateFromMetaforecastForm";
+
+export default async function CreateQuestionSetFromMetaforecastPage() {
+  const page = await loadTopBinaryQuestions();
+
+  return <CreateFromMetaforecastForm page={page} />;
+}
