+++ apps/hub/src/app/question-sets/create/page.tsx
@@ -6,12 +6,10 @@ import { FormProvider } from "react-hook-form";
 import { Button, TextFormField } from "@quri/ui";
 
 import { Card } from "@/components/ui/Card";
-import { H2 } from "@/components/ui/Headers";
-import { StyledLink } from "@/components/ui/StyledLink";
 import { createQuestionSetAction } from "@/evals/actions/createQuestionSetAction";
 import { QuestionsFieldArray } from "@/evals/components/QuestionsFieldArray";
 import { useSafeActionForm } from "@/lib/hooks/useSafeActionForm";
-import { questionSetRoute, questionSetsRoute } from "@/lib/routes";
+import { questionSetRoute } from "@/lib/routes";
 
 type FormShape = {
   name: string;
@@ -47,40 +45,31 @@ export default function CreateQuestionSetPage() {
       .questions.some((question) => question.description.trim() !== "");
 
   return (
-    <div>
-      <div className="mb-6 flex items-center justify-between">
-        <H2>Create New Question Set</H2>
-        <StyledLink href={questionSetsRoute()}>
-          ← Back to Question Sets
-        </StyledLink>
-      </div>
+    <Card theme="big">
+      <FormProvider {...form}>
+        <form onSubmit={onSubmit}>
+          <div className="mb-4">
+            <TextFormField
+              name="name"
+              label="Question Set Name"
+              placeholder="Enter a name for this question set"
+              rules={{ required: "Name is required" }}
+            />
+          </div>
 
-      <Card theme="big">
-        <FormProvider {...form}>
-          <form onSubmit={onSubmit}>
-            <div className="mb-4">
-              <TextFormField
-                name="name"
-                label="Question Set Name"
-                placeholder="Enter a name for this question set"
-                rules={{ required: "Name is required" }}
-              />
-            </div>
+          <QuestionsFieldArray name="questions" />
 
-            <QuestionsFieldArray name="questions" />
-
-            <div className="mt-6">
-              <Button
-                type="submit"
-                theme="primary"
-                disabled={!isValid || inFlight}
-              >
-                {inFlight ? "Creating..." : "Create Question Set"}
-              </Button>
-            </div>
-          </form>
-        </FormProvider>
-      </Card>
-    </div>
+          <div className="mt-6">
+            <Button
+              type="submit"
+              theme="primary"
+              disabled={!isValid || inFlight}
+            >
+              {inFlight ? "Creating..." : "Create Question Set"}
+            </Button>
+          </div>
+        </form>
+      </FormProvider>
+    </Card>
   );
 }
